스프링으로 답변형 게시판 만들기

+++ pro30/src/main/java/com/myspring/pro30/board/controller/BoardControllerImpl.java
@@ -1,6 +1,7 @@
 package com.myspring.pro30.board.controller;
 
 import java.io.File;
+import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -19,6 +20,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
@@ -26,6 +28,7 @@ import org.springframework.web.servlet.ModelAndView;
 
 import com.myspring.pro30.board.service.BoardService;
 import com.myspring.pro30.board.vo.ArticleVO;
+import com.myspring.pro30.board.vo.ImageVO;
 import com.myspring.pro30.member.vo.MemberVO;
 
 @Controller("boardController")
@@ -55,6 +58,7 @@ public class BoardControllerImpl implements BoardController{
 	@ResponseBody
 	public ResponseEntity addNewArticle(MultipartHttpServletRequest multipartRequest, HttpServletResponse response) throws Exception{
 		multipartRequest.setCharacterEncoding("utf-8");
+		String imageFileName = null;
 		//글 정보를 저정하기 위한 articleMap을 생성
 		Map<String, Object> articleMap = new HashMap<String, Object>();
 		Enumeration enu = multipartRequest.getParameterNames();
@@ -64,16 +68,27 @@ public class BoardControllerImpl implements BoardController{
 			String value=multipartRequest.getParameter(name);
 			articleMap.put(name, value);
 		}
-		//업로드한 이미지 파일 이름을 가져옴
-		String imageFileName = upload(multipartRequest);
+		//로그인 시 세션에 저장된 회원 정보에서 글쓴이 아이디를 얻어와서 Map에 저장
 		HttpSession session = multipartRequest.getSession();
 		MemberVO memberVO = (MemberVO) session.getAttribute("member"); //세션에 회원 정보 저장
 		//세선에 저장된 회원 정보로부터 회원 ID를 가져옴
 		String id = memberVO.getId();
 		//회원 ID, 이미지 파일 이름, 부모 글 번호를 articleMap에 저장
-		articleMap.put("id", id);
 		articleMap.put("parentNO", 0);
-		articleMap.put("imageFileName", imageFileName);
+		articleMap.put("id", id);
+		
+		//upload()를 호출해서 첨부한 이미지 파일 이름이 저장된 fileList을 받아옴
+		List<String> fileList = upload(multipartRequest);
+		List<ImageVO> imageFileList = new ArrayList<ImageVO>();
+		if(fileList != null && fileList.size() != 0) {
+			//전송된 이미지 정보를 ImageVO 객체의 속성에 차례대로 저장한 후 imageFileList에 다시 저장
+			for(String fileName : fileList) {
+				ImageVO imageVO = new ImageVO();
+				imageVO.setImageFileName(fileName);
+				imageFileList.add(imageVO);
+			}
+			articleMap.put("imageFileList", imageFileList);
+		}
 		
 		String message;
 		ResponseEntity resEnt = null;
@@ -84,10 +99,14 @@ public class BoardControllerImpl implements BoardController{
 			//글 정보가 저장된 articleMap을 Service 클래스의 addArticle() 메서드로 전달
 			int articleNO = boardService.addNewArticle(articleMap);
 			
-			if(imageFileName != null && imageFileName.length() != 0) {
-				File srcFile = new File(ARTICLE_IMAGE_REPO+"\\"+"temp"+"\\"+imageFileName);
-				File destDir = new File(ARTICLE_IMAGE_REPO+"\\"+articleNO);
-				FileUtils.moveFileToDirectory(srcFile, destDir, true);
+			if(imageFileList != null && imageFileList.size() != 0) {
+				//첨부한 이미지들을 for문으르 이용해 업로드함
+				for(ImageVO imageVO : imageFileList) {
+					imageFileName = imageVO.getImageFileName();
+					File srcFile = new File(ARTICLE_IMAGE_REPO+"\\"+"temp"+"\\"+imageFileName);
+					File destDir = new File(ARTICLE_IMAGE_REPO+"\\"+articleNO);
+					FileUtils.moveFileToDirectory(srcFile, destDir, true);
+				}
 			}//글 정보를 추가한 후 업로드한 이미지 파일을 글 번호로 명명한 폴더로 이동
 			
 			//새 글을 추가한 후 메시지를 전달
@@ -97,8 +116,13 @@ public class BoardControllerImpl implements BoardController{
 			message += " </script>";
 			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
 		}catch(Exception e) {
-			File srcFile = new File(ARTICLE_IMAGE_REPO+"\\"+"temp"+"\\"+imageFileName);
-			srcFile.delete();
+			if(imageFileList != null && imageFileList.size() != 0) {
+				for(ImageVO imageVO:imageFileList) {
+					imageFileName = imageVO.getImageFileName();
+					File srcFile = new File(ARTICLE_IMAGE_REPO+"\\"+"temp"+"\\"+imageFileName);
+					srcFile.delete();
+				}
+			}
 			
 			//오류 발생 시 오류 메시지 전달
 			message = "<script>";
@@ -120,23 +144,136 @@ public class BoardControllerImpl implements BoardController{
 		return mav;
 	}
 	
-	//한 개 이미지 업로드한 파일 이름을 얻은 후 반환
-	private String upload(MultipartHttpServletRequest multipartRequest) throws Exception{
+	//글 상세창을 나타냄
+	@Override
+	@RequestMapping(value="/board/viewArticle.do" ,method = RequestMethod.GET)
+	public ModelAndView viewArticle(@RequestParam("articleNO") int articleNO,
+                                    HttpServletRequest request, HttpServletResponse response) throws Exception{
+		String viewName = (String)request.getAttribute("viewName");
+		Map articleMap = boardService.viewArticle(articleNO);
+		ModelAndView mav = new ModelAndView();
+		mav.setViewName(viewName);
+		mav.addObject("articleMap", articleMap);
+		return mav;
+	}
+	
+	//글 수정하기
+	@Override
+	@RequestMapping(value="/board/modArticle.do", method=RequestMethod.POST)
+	@ResponseBody
+	public ResponseEntity modArticle(MultipartHttpServletRequest multipartRequest, HttpServletResponse response) throws Exception{
+		multipartRequest.setCharacterEncoding("utf-8");
 		String imageFileName = null;
+		
+		//글 정보를 저장하기 위해 articleMap 생성
+		Map<String, Object> articleMap = new HashMap<String, Object>();
+		Enumeration enu = multipartRequest.getParameterNames();
+		//글 수정창에서 전송된 글 정보를 key/value로 저장
+		while(enu.hasMoreElements()) {
+			String name = (String)enu.nextElement();
+			String value = multipartRequest.getParameter(name);
+			articleMap.put(name, value);
+		}
+		HttpSession session = multipartRequest.getSession();
+		
+		List<String> fileList = upload(multipartRequest);
+		List<ImageVO> imageFileList = new ArrayList<ImageVO>();
+		
+		articleMap.put("imageFileList", imageFileList);
+		
+		String articleNO = (String)articleMap.get("articleNO"); //articleMap의 글 번호를 가져옴
+		String message;
+		ResponseEntity resEnt = null;
+		HttpHeaders responseHeaders = new HttpHeaders();
+		responseHeaders.add("Content-Type", "text/html; charset=utf-8");
+		try {
+			boardService.modArticle(articleMap);
+			if(imageFileList != null && imageFileList.size() != 0) {
+				for(ImageVO imageVO:imageFileList) {
+					imageFileName = imageVO.getImageFileName();
+					//새로 첨부한 파일을 폴더로 이동
+					File srcFile = new File(ARTICLE_IMAGE_REPO+"\\"+"temp"+"\\"+imageFileName);
+					File destDir = new File(ARTICLE_IMAGE_REPO+"\\"+articleNO);
+					FileUtils.moveFileToDirectory(srcFile, destDir, true);
+					
+					//기존 파일 삭제
+					String originalFileName = (String)articleMap.get("originalFileName");
+					File oldFile = new File(ARTICLE_IMAGE_REPO+"\\"+articleNO+"\\"+originalFileName);
+					oldFile.delete();
+				}
+			}
+			message = "<script>";
+			message += " alert('글을 수정합니다.');";
+			message += "location.href='"+multipartRequest.getContextPath()+"/board/viewArticle.do?articleNO="+articleNO+"';";
+			message += "</script>";
+			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
+		}catch(Exception e) {
+			if(imageFileList != null && imageFileList.size() != 0) {
+				for(ImageVO imageVO : imageFileList) {
+					imageFileName = imageVO.getImageFileName();
+					File srcFile = new File(ARTICLE_IMAGE_REPO+"\\"+"temp"+"\\"+imageFileName);
+					srcFile.delete();
+				}
+			}
+			message = "<script>";
+			message += " alert('오류가 발생했습니다. 다시 수정해주세요.');";
+			message += " loction.href='"+multipartRequest.getContextPath()+"/board/viewArticle.do?articleNO="+articleNO+"';";
+			message += "</script>";
+			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
+		}
+		return resEnt;
+	}
+	
+	//해당 글 삭제하기
+	@Override
+	@RequestMapping(value="/board/removeArticle.do", method=RequestMethod.POST)
+	@ResponseBody
+	public ResponseEntity removeArticle(@RequestParam("articleNO") int articleNO, HttpServletRequest request, HttpServletResponse response) throws Exception{
+		response.setContentType("text/html; charset=utf-8");
+		String message;
+		ResponseEntity resEnt = null;
+		HttpHeaders responseHeaders = new HttpHeaders();
+		responseHeaders.add("Content-Type", "text/html; charset=utf-8");
+		try {
+			boardService.removeArticle(articleNO);
+			File desDir = new File(ARTICLE_IMAGE_REPO+"\\"+articleNO);
+			FileUtils.deleteDirectory(desDir);
+			
+			message = "<script>";
+			message += " alert('글을 삭제했습니다.');";
+			message += " location.href='"+request.getContextPath()+"/board/listArticles.do';";
+			message += "</script>";
+			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
+		}catch(Exception e) {
+			message ="<script>";
+			message +=" alert('작업 중 오류가 발생했습니다. 다시 시도해 주세요.');";
+			message +=" location.href='"+request.getContextPath()+"/board/listArticles.do';";
+			message += " </script>";
+			resEnt = new ResponseEntity(message, responseHeaders, HttpStatus.CREATED);
+			e.printStackTrace();
+		}
+		return resEnt;
+	}
+	
+	
+	//한 개 이미지 업로드한 파일 이름을 얻은 후 반환
+	private List<String> upload(MultipartHttpServletRequest multipartRequest) throws Exception{
+		List<String> fileList = new ArrayList<String>();
 		Iterator<String> fileNames = multipartRequest.getFileNames();
 		while(fileNames.hasNext()) {
 			String fileName = fileNames.next();
 			MultipartFile mFile = multipartRequest.getFile(fileName);
-			imageFileName = mFile.getOriginalFilename();
+			String originalFileName = mFile.getOriginalFilename();
+			fileList.add(originalFileName); //첨부한 이미지 파일의 이름들을 차례대로 저장함
 			File file = new File(ARTICLE_IMAGE_REPO+"\\"+"temp"+"\\"+fileName);
 			if(mFile.getSize()!=0) { //파일 null 체크
 				if(!file.exists()) { //경로상에 파일이 존재하지 않을 경우
 					file.getParentFile().mkdirs(); //경로에 해당하는 디렉토리들을 생성
-					mFile.transferTo(new File(ARTICLE_IMAGE_REPO+"\\"+"temp"+"\\"+imageFileName)); //임시로 저장된 multipartFile를 실제 파일로 전송
+					mFile.transferTo(new File(ARTICLE_IMAGE_REPO+"\\"+"temp"+"\\"+originalFileName)); //임시로 저장된 multipartFile를 실제 파일로 전송
 				}
 			}
 		}
-		return imageFileName;
+		return fileList;
 	}
 	
 }
